Turn logcollector into an operator sub-command
This splits the operator binary into two sub-commands: `operator` and
`logcollector`, and removes the building/pushing of the logcollector images.
References to the logcollector image are replaced by the operator image, and
run with an invocation of the logcollector sub-command.
