+++ Makefile
@@ -179,12 +179,12 @@ endif
 	@echo "Running e2e tests"
 	@echo "WARNING: This will temporarily modify deploy/operator.yaml"
 	@echo "Replacing workload references in deploy/operator.yaml"
-	@sed -i 's%$(IMAGE_REPO)/$(LOG_COLLECTOR):latest%$(LOGCOLLECTOR_IMAGE_PATH)%' deploy/operator.yaml
 	@sed -i 's%$(IMAGE_REPO)/$(AIDE):latest%$(AIDE_IMAGE_PATH)%' deploy/operator.yaml
+	@sed -i 's%$(IMAGE_REPO)/$(APP_NAME):latest%$(OPERATOR_IMAGE_PATH)%' deploy/operator.yaml
 	unset GOFLAGS && $(GOPATH)/bin/operator-sdk test local ./tests/e2e --skip-cleanup-error --image "$(OPERATOR_IMAGE_PATH)" --namespace "$(NAMESPACE)" --go-test-flags "$(E2E_GO_TEST_FLAGS)"
 	@echo "Restoring image references in deploy/operator.yaml"
-	@sed -i 's%$(LOGCOLLECTOR_IMAGE_PATH)%$(IMAGE_REPO)/$(LOG_COLLECTOR):latest%' deploy/operator.yaml
 	@sed -i 's%$(AIDE_IMAGE_PATH)%$(IMAGE_REPO)/$(AIDE):latest%' deploy/operator.yaml
+	@sed -i 's%$(OPERATOR_IMAGE_PATH)%$(IMAGE_REPO)/$(APP_NAME):latest%' deploy/operator.yaml
 
 # If IMAGE_FORMAT is not defined, it means that we're not running on CI, so we
 # probably want to push the file-integrity-operator image to the cluster we're
@@ -209,8 +209,6 @@ image-to-cluster:
 	@echo "Skipping image-to-cluster target."
 	$(eval component = $(APP_NAME))
 	$(eval OPERATOR_IMAGE_PATH = $(IMAGE_FORMAT))
-	$(eval component = file-integrity-logcollector)
-	$(eval LOGCOLLECTOR_IMAGE_PATH = $(IMAGE_FORMAT))
 	# TODO(jaosorior): Add AIDE image to release repo and subsequently add it here.
 else ifeq ($(E2E_USE_DEFAULT_IMAGES), true)
 image-to-cluster:
@@ -219,7 +217,7 @@ else ifeq ($(E2E_SKIP_CONTAINER_PUSH), true)
 image-to-cluster:
 	@echo "E2E_SKIP_CONTAINER_PUSH variable detected. Using previously pushed images."
 	$(eval OPERATOR_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(APP_NAME):$(TAG))
-	$(eval LOGCOLLECTOR_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(LOG_COLLECTOR):$(TAG))
+	$(eval LOGCOLLECTOR_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(APP_NAME):$(TAG))
 	$(eval AIDE_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(AIDE):$(TAG))
 else
 image-to-cluster: namespace openshift-user image
@@ -230,12 +228,11 @@ image-to-cluster: namespace openshift-user image
 	IMAGE_REGISTRY_HOST=$$(oc get route default-route -n openshift-image-registry --template='{{ .spec.host }}'); \
 		$(RUNTIME) login --tls-verify=false -u $(OPENSHIFT_USER) -p $(shell oc whoami -t) $${IMAGE_REGISTRY_HOST}; \
 		$(RUNTIME) push --tls-verify=false $(OPERATOR_IMAGE_PATH):$(TAG) $${IMAGE_REGISTRY_HOST}/$(NAMESPACE)/$(APP_NAME):$(TAG); \
-		$(RUNTIME) push --tls-verify=false $(LOGCOLLECTOR_IMAGE_PATH):$(TAG) $${IMAGE_REGISTRY_HOST}/$(NAMESPACE)/$(LOG_COLLECTOR):$(TAG); \
 		$(RUNTIME) push --tls-verify=false $(AIDE_IMAGE_PATH):$(TAG) $${IMAGE_REGISTRY_HOST}/$(NAMESPACE)/$(AIDE):$(TAG)
 	@echo "Removing the route from the image registry"
 	@oc patch configs.imageregistry.operator.openshift.io/cluster --patch '{"spec":{"defaultRoute":false}}' --type=merge
 	$(eval OPERATOR_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(APP_NAME):$(TAG))
-	$(eval LOGCOLLECTOR_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(LOG_COLLECTOR):$(TAG))
+	$(eval LOGCOLLECTOR_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(APP_NAME):$(TAG))
 	$(eval AIDE_IMAGE_PATH = image-registry.openshift-image-registry.svc:5000/$(NAMESPACE)/$(AIDE):$(TAG))
 endif
 
@@ -255,7 +252,6 @@ endif
 .PHONY: push
 push: image
 	$(RUNTIME) push $(OPERATOR_IMAGE_PATH):$(TAG)
-	$(RUNTIME) push $(LOGCOLLECTOR_IMAGE_PATH):$(TAG)
 	$(RUNTIME) push $(AIDE_IMAGE_PATH):$(TAG)
 
 .PHONY: publish
@@ -285,13 +281,11 @@ package-version-to-tag: check-package-version
 .PHONY: release-tag-image
 release-tag-image: package-version-to-tag
 	@echo "Temporarily overriding image tags in deploy/operator.yaml"
-	@sed -i 's%$(IMAGE_REPO)/$(LOG_COLLECTOR):latest%$(LOGCOLLECTOR_IMAGE_PATH):$(TAG)%' deploy/operator.yaml
 	@sed -i 's%$(IMAGE_REPO)/$(AIDE):latest%$(AIDE_IMAGE_PATH):$(TAG)%' deploy/operator.yaml
 
 .PHONY: undo-deploy-tag-image
 undo-deploy-tag-image: package-version-to-tag
 	@echo "Restoring image tags in deploy/operator.yaml"
-	@sed -i 's%$(LOGCOLLECTOR_IMAGE_PATH):$(TAG)%$(IMAGE_REPO)/$(LOG_COLLECTOR):latest%' deploy/operator.yaml
 	@sed -i 's%$(AIDE_IMAGE_PATH):$(TAG)%$(IMAGE_REPO)/$(AIDE):latest%' deploy/operator.yaml
 
 .PHONY: git-release
