+++ go.mod
@@ -13,7 +13,6 @@ require (
 	github.com/pkg/errors v0.8.1
 	github.com/securego/gosec v0.0.0-20200330112059-e030aa4f768b
 	github.com/spf13/cobra v0.0.5
-	github.com/spf13/pflag v1.0.5
 	k8s.io/api v0.17.4
 	k8s.io/apimachinery v0.17.4
 	k8s.io/client-go v12.0.0+incompatible
